Update README.md
I have successfully created a portfolio website using HTML and CSS and Javascript as part of a project. This project helped me improve my web development skills and learn new technologies. I used HTML to create the content and CSS to design the layout, colors, and fonts. To improve the website, I made it dynamic and mobile friendly with flexbox and javascript